Added and fixed the deletion of grids when above the max amount of non active grids, Terrain Placer Script is now obsolete

+++ Assets/Scripts/TerrainPlacer.cs
@@ -18,13 +18,13 @@ public class TerrainPlacer : MonoBehaviour
         Vector3 newPos = transform.position;
         newPos.y = 0;
         
-        newPos -= new Vector3(xSize/2, 0, 0);
+        newPos -= new Vector3((float)xSize/2, 0, 0);
         placeTerrain(newPos);
         
-        newPos -= new Vector3(xSize, 0, 0);
+        newPos -= new Vector3((float)xSize, 0, 0);
         placeTerrain(newPos);
         
-        newPos += new Vector3(xSize*2, 0, 0);
+        newPos += new Vector3((float)xSize*2, 0, 0);
         placeTerrain(newPos);
 
     }
